+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Properly make newly inserted location visible (scroll to row), specifically when location table is sorted by ID in descending order
 - Do not reset backup period to "Never" when setting the backup path fails
 - Correctly name the file selection filter of the Sky Dolly logbook location import plugin
+- Properly enable/disable pitch, bank, true heading and indicated airspeed spinboxes based on location selection
 
 ## 0.19.0
 Sky Dolly **"Jaunty Jodel"** carries the Olympic spirit into the world: over 200 new locations - each country taking part in the Olympic Games in Paris 2024 - have been added as location presets. Please enjoy the diversity and beauty of our little planet - happy flying!